Completed Tutorial
Completed Brackey's Godot Tutorial, however I am not done with this project yet. I am aiming to make a larger map, add in new gameplay elements and QOL changes that will make the game far far more refined, as currently it is not a very good playable game. Areas are unreachable and dead ends, there are very few coins, and the text is just placed in a random spot.